wip(frontend): replace searcharea component by simplified and general component everywhere
